* apache/master: (113 commits)
  [SPARK-8196][SQL] Fix null handling & documentation for next_day.
  [SPARK-9373][SQL] follow up for StructType support in Tungsten projection.
  [SPARK-9402][SQL] Remove CodegenFallback from Abs / FormatNumber.
  [SPARK-8919] [DOCUMENTATION, MLLIB] Added @since tags to mllib.recommendation
  [EC2] Cosmetic fix for usage of spark-ec2 --ebs-vol-num option
  [SPARK-9394][SQL] Handle parentheses in CodeFormatter.
  Closes #6836 since Round has already been implemented.
  [SPARK-9335] [STREAMING] [TESTS] Make sure the test stream is deleted in KinesisBackedBlockRDDSuite
  [MINOR] [SQL] Support mutable expression unit test with codegen projection
  [SPARK-9373][SQL] Support StructType in Tungsten projection
  [SPARK-8828] [SQL] Revert SPARK-5680
  Fixed a test failure.
  [SPARK-9395][SQL] Create a SpecializedGetters interface to track all the specialized getters.
  [SPARK-8195] [SPARK-8196] [SQL] udf next_day last_day
  [SPARK-8882] [STREAMING] Add a new Receiver scheduling mechanism
  [SPARK-9386] [SQL] Feature flag for metastore partition pruning
  [SPARK-9230] [ML] Support StringType features in RFormula
  [SPARK-9385] [PYSPARK] Enable PEP8 but disable installing pylint.
  [SPARK-4352] [YARN] [WIP] Incorporate locality preferences in dynamic allocation requests
  [SPARK-9385] [HOT-FIX] [PYSPARK] Comment out Python style check
  ...

+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/expressions/BitwiseFunctionsSuite.scala
@@ -30,8 +30,9 @@ class BitwiseFunctionsSuite extends SparkFunSuite with ExpressionEvalHelper {
       checkEvaluation(expr, expected)
     }
 
-    check(1.toByte, ~1.toByte)
-    check(1000.toShort, ~1000.toShort)
+    // Need the extra toByte even though IntelliJ thought it's not needed.
+    check(1.toByte, (~1.toByte).toByte)
+    check(1000.toShort, (~1000.toShort).toShort)
     check(1000000, ~1000000)
     check(123456789123L, ~123456789123L)
 
@@ -45,8 +46,9 @@ class BitwiseFunctionsSuite extends SparkFunSuite with ExpressionEvalHelper {
       checkEvaluation(expr, expected)
     }
 
-    check(1.toByte, 2.toByte, 1.toByte & 2.toByte)
-    check(1000.toShort, 2.toShort, 1000.toShort & 2.toShort)
+    // Need the extra toByte even though IntelliJ thought it's not needed.
+    check(1.toByte, 2.toByte, (1.toByte & 2.toByte).toByte)
+    check(1000.toShort, 2.toShort, (1000.toShort & 2.toShort).toShort)
     check(1000000, 4, 1000000 & 4)
     check(123456789123L, 5L, 123456789123L & 5L)
 
@@ -63,8 +65,9 @@ class BitwiseFunctionsSuite extends SparkFunSuite with ExpressionEvalHelper {
       checkEvaluation(expr, expected)
     }
 
-    check(1.toByte, 2.toByte, 1.toByte | 2.toByte)
-    check(1000.toShort, 2.toShort, 1000.toShort | 2.toShort)
+    // Need the extra toByte even though IntelliJ thought it's not needed.
+    check(1.toByte, 2.toByte, (1.toByte | 2.toByte).toByte)
+    check(1000.toShort, 2.toShort, (1000.toShort | 2.toShort).toShort)
     check(1000000, 4, 1000000 | 4)
     check(123456789123L, 5L, 123456789123L | 5L)
 
@@ -81,8 +84,9 @@ class BitwiseFunctionsSuite extends SparkFunSuite with ExpressionEvalHelper {
       checkEvaluation(expr, expected)
     }
 
-    check(1.toByte, 2.toByte, 1.toByte ^ 2.toByte)
-    check(1000.toShort, 2.toShort, 1000.toShort ^ 2.toShort)
+    // Need the extra toByte even though IntelliJ thought it's not needed.
+    check(1.toByte, 2.toByte, (1.toByte ^ 2.toByte).toByte)
+    check(1000.toShort, 2.toShort, (1000.toShort ^ 2.toShort).toShort)
     check(1000000, 4, 1000000 ^ 4)
     check(123456789123L, 5L, 123456789123L ^ 5L)
